refactor: move ChainId to dedicated file
