+++ packages/arb-token-bridge-ui/src/components/TransferPanel/EstimatedGas.tsx
@@ -3,7 +3,8 @@ import { useMemo } from 'react'
 import { twMerge } from 'tailwind-merge'
 
 import { useAppState } from '../../state'
-import { ChainId, getNetworkName, isNetwork } from '../../util/networks'
+import { getNetworkName, isNetwork } from '../../util/networks'
+import { ChainId } from '../../types/ChainId'
 import { Tooltip } from '../common/Tooltip'
 import { formatAmount } from '../../util/NumberUtils'
 import { useNativeCurrency } from '../../hooks/useNativeCurrency'
