+++ packages/arb-token-bridge-ui/src/generateOpenGraphImages.tsx
@@ -2,7 +2,8 @@ import React from 'react'
 import satori, { Font } from 'satori'
 import sharp from 'sharp'
 import fs from 'fs'
-import { ChainId, isNetwork } from './util/networks'
+import { isNetwork } from './util/networks'
+import { ChainId } from './types/ChainId'
 import { getBridgeUiConfigForChain } from './util/bridgeUiConfig'
 import { orbitMainnets, orbitTestnets } from './util/orbitChainsList'
