+++ packages/arb-token-bridge-ui/src/hooks/__tests__/useArbQueryParams.test.ts
@@ -2,7 +2,8 @@
  * @jest-environment jsdom
  */
 import { registerCustomArbitrumNetwork } from '@arbitrum/sdk'
-import { ChainId, customChainLocalStorageKey } from '../../util/networks'
+import { customChainLocalStorageKey } from '../../util/networks'
+import { ChainId } from '../../types/ChainId'
 import { AmountQueryParam, ChainParam } from '../useArbQueryParams'
 import { createMockOrbitChain } from './helpers'
