+++ packages/arb-token-bridge-ui/src/types/ChainQueryParam.ts
@@ -1,10 +1,10 @@
 import { Chain } from 'wagmi'
 import * as chains from 'wagmi/chains'
 import {
-  ChainId,
   getCustomChainFromLocalStorageById,
   getSupportedChainIds
 } from '../util/networks'
+import { ChainId } from '../types/ChainId'
 import * as customChains from '../util/wagmi/wagmiAdditionalNetworks'
 import { getOrbitChains, orbitChains } from '../util/orbitChainsList'
 import { chainToWagmiChain } from '../util/wagmi/wagmiAdditionalNetworks'
