+++ packages/arb-token-bridge-ui/src/util/TokenTeleportEnabledUtils.ts
@@ -1,4 +1,4 @@
-import { ChainId } from './networks'
+import { ChainId } from '../types/ChainId'
 
 export type TeleportEnabledToken = {
   symbol: string
