+++ packages/arb-token-bridge-ui/src/util/TokenUtils.ts
@@ -11,7 +11,8 @@ import {
 import { ERC20__factory } from '@arbitrum/sdk/dist/lib/abi/factories/ERC20__factory'
 
 import { CommonAddress } from './CommonAddressUtils'
-import { ChainId, isNetwork } from './networks'
+import { isNetwork } from './networks'
+import { ChainId } from '../types/ChainId'
 import { defaultErc20Decimals } from '../defaults'
 import { ERC20BridgeToken, TokenType } from '../hooks/arbTokenBridge.types'
 import { getBridger, getChainIdFromProvider } from '../token-bridge-sdk/utils'
