+++ packages/arb-token-bridge-ui/src/util/__tests__/networks.test.ts
@@ -1,11 +1,11 @@
 import { registerCustomArbitrumNetwork } from '@arbitrum/sdk'
 
 import {
-  ChainId,
   getBlockNumberReferenceChainIdByChainId,
   getDestinationChainIds,
   getSupportedChainIds
 } from '../networks'
+import { ChainId } from '../../types/ChainId'
 import { orbitTestnets } from '../orbitChainsList'
 
 const xaiTestnetChainId = 37714555429
